% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TM_player.R
\name{get_TM_player}
\alias{get_TM_player}
\title{get_TM_player}
\usage{
get_TM_player(player_url, user_agent, raw = FALSE)
}
\arguments{
\item{user_agent}{A character introducing yourself to the website - REQUIRED.}

\item{raw}{Logical: if TRUE, returns the unprocessed table. if FALSE, returns a processed tibble - defaults to FALSE}

\item{playerurl}{A valid transfermarkt url. Navigate to a player's page, and then copy the url - REQUIRED.}
}
\description{
A webscraping function for the website 'https://www.transfermarkt.com/'.
Returns a preprocessed tibble of a player.
}
\examples{
dusan_vlahovic <- get_TM_player(player_url = 'https://www.transfermarkt.co.uk/du-scaron-an-vlahović/profil/spieler/357498', user_agent = 'John Smith personal project')
}
\keyword{Football,}
\keyword{Soccer}
\keyword{Webscraping,}
