% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TM_squad.R
\name{get_TM_squad}
\alias{get_TM_squad}
\title{get_TM_squad}
\usage{
get_TM_squad(squad_url, user_agent, raw = FALSE, year = 2022)
}
\arguments{
\item{user_agent}{A character introducing yourself to the website - REQUIRED.}

\item{raw}{Logical: if TRUE, returns the unprocessed table. if FALSE, returns a processed tibble - defaults to FALSE}

\item{year}{Numeric: specify the year for which you would like the data}

\item{url}{A valid ()detailed) transfermarkt url. Navigate to a club page, click on the 'detailed' tab, and then paste the url - REQUIRED.}
}
\description{
A webscraping function for the website 'https://www.transfermarkt.com/'.
Returns a preprocessed tibble of any squad.
}
\examples{
Juventus squad <- get_TM_squad(squad_url = 'https://www.transfermarkt.co.uk/juventus-turin/startseite/verein/506', user_agent = 'John Smith personal project')
Blackburn_Rovers_squad <- get_TM_squad(squad_url = 'https://www.transfermarkt.co.uk/blackburn-rovers/startseite/verein/164', user_agent = 'John Smith personal project')
}
\keyword{Football,}
\keyword{Soccer}
\keyword{Webscraping,}
